Update IrcDotNet to work with Twitch.tv's IRC server
Twitch's IRC server does not work *exactly* like a real IRC server.
Primarily, it does not support WHO requests, and it does not give JOIN
and PART messages.  IrcDotNet takes a dependency on a user having
JOIN'ed a channel before being able to set the mode for the user (which
is a reasonable assumption, but not true in the case of twitch).

Additionally, some messages twitch sends are not in the standard format,
so we have to be careful when parsing the data from twitch.

This change allows IrcDotNet to connect to twitch.tv.  However, it also
makes it so that it may be incompatible with other IRC servers.